requirements.txt
icmplib
matplotlib
pydub
chat_exporter
youtube_search_python
pypresence
aiohttp
polaroid
glitch-this
dateparser
gTTS
colorama
psutil
async_cse
unshortenit
pycountry
humanize
numpy
urllib3
wordcloud
discord.py-self
pyPrivnote
opencv_python
Pillow
akinator.py
youtube_dl
